Handle static library, selection of compiler, and etc
Ctypesgen doesn't support static libraries. If user passes static
library, automaticaly create a shared library for it and use it for
python bindings.

By default the compiler for preprocessing is set to CMAKE_CXX_COMPILER
but if the target uses CMAKE_C_COMPILER it will set to that.

Adding ctypesgen_def.h header file for defining things. It is useful for
handling things that ctypesgen doesn't support. For example msvc has
built-in sized arguments named __int64 and so. When ctypes gen
encounters any of symbol which uses them it yields with error and skips
creating python bindings for it.